refactor: モデル定義を別ファイルに切り出した
