+++ app/models/api_models.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel
+class Request(BaseModel):
+    location: str
+    prompt:  str
+
+class Response(BaseModel):
+    id: int
+    name: str
+    long: float
+    lat: float
+    area: str
+    category: str
